requirements.txt
dj-database-url==0.5.0
Django==2.2
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.9.2
django-heroku==0.3.1
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.6